# api/config/openapi.yaml
openapi: 3.0.0
info:
  title: API de Certificação de Documentos
  version: 1.0.0
  description: API para certificar documentos via blockchain
servers:
  - url: http://localhost:3000
    description: Servidor local
paths:
  /auth/login:
    post:
      summary: Loga no sistema utilizando dados da carteira
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                address:
                  type: string
                  description: Endereço da carteira Ethereum
                  example: "0x1234567890abcdef"
                signature:
                  type: string
                  description: Assinatura gerada pela chave privada
                  example: "0xabcdef1234567890"
                nonce:
                  type: string
                  description: Nonce gerado pelo backend para garantir unicidade da assinatura
                  example: "xyz1234"
      responses:
        '200':
          description: Login bem-sucedido
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Login bem-sucedido
        '400':
          description: Falha na autenticação
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Falha na autenticação
  /auth/nonce:
    get:
      summary: Gera o nonce para fazer login
      tags:
        - Auth
      parameters:
        - in: query
          name: address
          required: true
          schema:
            type: string
          description: Endereço da carteira Ethereum, utilizado para gerar o nonce
      responses:
        '200':
          description: Nonce gerado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  nonce:
                    type: string
                    example: "xyz1234"  # Aqui, o exemplo de nonce gerado
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Endereço inválido ou faltando"
...